% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phorest.R
\name{phorest}
\alias{phorest}
\title{plots multiple phenotypic forests plots for cross-comparison.}
\usage{
phorest(associations, columnNames = NULL, phenotypeGroups = NULL,
  unassignedGroup = "unassigned", transparency = 0.6,
  legendTransparency = FALSE, connectingLines = FALSE,
  connectingLinesColour = "#707070", legendSNP = TRUE, xLabel = NULL,
  yLabel = NULL, title = "", warningOn = TRUE, verticalLine = NULL,
  largeSNPSet = FALSE, SNPGroups = NULL, aggregatingFunction = function(x)
  median(x, na.rm = TRUE), aggregatingLowerBound = function(x) quantile(x,
  0.25, na.rm = TRUE), aggregatingUpperBound = function(x) quantile(x, 0.75,
  na.rm = TRUE), transparencyAggregate = 0.5)
}
\arguments{
\item{associations}{data frame with at least columns: phenotype, value ,lowerBound,upperBound and SNP, or phenotype, value and SNP if largeSNPSet is TRUE. phenotypeGroup and SNPGroup can also be defined here.}

\item{columnNames}{user-defined columns names}

\item{phenotypeGroups}{list of phenotypes to be grouped together.}

\item{unassignedGroup}{default group name.}

\item{transparency}{transparency of the error bar. A number between 0 and 1. Can be a vector of length nrow(associations).}

\item{legendTransparency}{logical. Whether to show the legend of transparency.}

\item{connectingLines}{logical. Whether to show connecting lines for each SNP.}

\item{connectingLinesColour}{colour of the connecting lines.}

\item{legendSNP}{logical. Whether to show the legend for SNPs.}

\item{xLabel}{label on the x-axis}

\item{yLabel}{label on the y-axis}

\item{title}{plot title}

\item{warningOn}{whether warnings are shown.}

\item{verticalLine}{x intercepts for vertical lines}

\item{largeSNPSet}{logical. Whether to show an aggregate plot.}

\item{SNPGroups}{list of SNPs to be grouped together}

\item{aggregatingFunction}{function to use for aggregating the values of each group of SNPs}

\item{aggregatingLowerBound}{function to use for aggregating the values of each group of SNPs and get a lower bound}

\item{aggregatingUpperBound}{function to use for aggregating the values of each group of SNPs and get an upper bound}

\item{transparencyAggregate}{transparency of the connecting lines}
}
\value{
a ggplot object
}
\description{
The plot is produced from a data frame containing the phenotypes, phenotype groups, SNPs, some values and possibly some upper and lower bounds.
By default, \code{phorest()} expects the following columns: phenotype,phenotypeGroup,SNP,SNPGroup,value,lowerBound and upperBound.
But this can be over-ridden with the columnNames, phenotypeGroups and SNPGroups arguments.
SNP are assumed to be of the form 'rsXXXXX'; this is used for ordering labels in the legend.ddddddd
}
\examples{
set.seed(42)

nPhenotype <- 17
nSNP <- 7
nPhenotypeGroups <- 3

randomWord<-function(n, nLetters = 5)
  replicate(n,paste(sample(letters, nLetters, replace = TRUE),sep = '', collapse=''))

df <- data.frame(
  phenotype      = rep(randomWord(nPhenotype),1,each=nSNP),
  value          = rep(1:nSNP,nPhenotype) + rnorm(nSNP*nPhenotype,mean=0,sd=1),
  lowerBound     = runif(nPhenotype*nSNP,min=0.0,max=0.1),
  upperBound     = runif(nPhenotype*nSNP,min=0.0,max=0.1),
  phenotypeGroup = rep(sample(toupper(randomWord(nPhenotypeGroups)),nPhenotype,replace=TRUE),1,each=nSNP),
  SNP            = paste('rs',rep(sample(100000,nSNP), nPhenotype), sep = '')
)

df<-within(df,{
  lowerBound <- value - lowerBound
  upperBound <- value + upperBound}
)

print(head(df))
p <- phorest(df, connectingLines = TRUE)
print(p)
}
\author{
Christophe Ladroue
}

