% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/polarHistogram.R
\name{polarHistogram}
\alias{polarHistogram}
\title{builds many histograms and arranges them around a circle to save space.}
\usage{
polarHistogram(df, family = NULL, columnNames = NULL, binSize = 1,
  spaceItem = 0.2, spaceFamily = 1.2, innerRadius = 0.3,
  outerRadius = 1, guides = c(10, 20, 40, 80), alphaStart = -0.3,
  circleProportion = 0.8, direction = "inwards", familyLabels = FALSE,
  normalised = TRUE)
}
\arguments{
\item{df}{a data frame containing the data}

\item{family}{list for defining families}

\item{columnNames}{user-defined columns names}

\item{binSize}{width of the bin. Should probably be left as 1, as other parameters are relative to it.}

\item{spaceItem}{space between bins}

\item{spaceFamily}{space between families}

\item{innerRadius}{radius of inner circle}

\item{outerRadius}{radius of outer circle. Should probably be left as 1, as other parameters are relative to it.}

\item{guides}{a vector with percentages to use for the white guide lines}

\item{alphaStart}{offset from 12 o'clock in radians}

\item{circleProportion}{proportion of the circle to cover}

\item{direction}{either of "inwards" or "outwards". Whether the increasing count goes from or to the centre.}

\item{familyLabels}{logical. Whether to show family names}

\item{normalised}{logical.}
}
\value{
a ggplot object
}
\description{
The data frame is expected to have at least four columns: family, item, score and value.
The three first columns are categorical variables, the fourth column contains non-negative values.
See the vignette for an example.
Each bar represents the proportion of scores for one item. Items can be grouped by families.
The resulting graph can be busy and might be better off saved as a pdf, with ggsave("myGraph.pdf").
}
\examples{
set.seed(42)
nFamily <- 20
nItemPerFamily <- sample(1:6,nFamily,replace=TRUE)
nValues <- 3
randomWord <- function(n,nLetters=5)
  replicate(n, paste(sample(letters, nLetters, replace = TRUE), sep = '', collapse = ''))

df <- data.frame(
  family = rep( randomWord(nFamily), times = nValues * nItemPerFamily),
  item   = rep( randomWord(sum(nItemPerFamily), 3), each = nValues ),
  score  = rep( paste0("V",1:nValues), times = sum(nItemPerFamily)),
  value  = runif( sum(nItemPerFamily * nValues)))
print(head(df))
p <- polarHistogram(df, familyLabels = FALSE)
print(p)
}
\author{
Christophe Ladroue
}

